examples/semantic-code-search/requirements.txt
flask==3.0.2
flask-cors==4.0.0
gitpython==3.1.42
sentence-transformers==2.5.1
chromadb==0.4.24
python-dotenv==1.0.1
llama-stack==0.1.0
llama-stack-client==0.1.0
llama-models==0.1.0
together==1.3.11
tree-sitter==0.21.0